Add multi-platform support logic
Compiled libraries annoyingly have different prefixes
and extensions depending on the platform you're on.
This commit adds the logic needed to ensure that
imgal_java finds the native imgal_c cdylib on all
platforms.

+++ imgal_java/src/main/java/org/imgal/AbstractNativeLibrary.java
@@ -15,23 +15,53 @@ import java.nio.file.StandardCopyOption;
  * @author Edward Evans
  */
 public abstract class AbstractNativeLibrary {
-	// TODO support macOS and Windows style native libs
-	private static String libPath = "/native/libimgal_c.so";
+	private static final String libName = "imgal_c";
+	private static final String libPath;
+	private static final String libPrefix;
+	private static final String libExtension;
+	private static final String fileName;
 	public static final SymbolLookup libLookup;
 	public static final Linker linker = Linker.nativeLinker();
 
-	// copy the Rust library from resources and then load it (for SymbolLookup)
+
+	// determine the platform specific library path
+	static {
+		String os = System.getProperty("os.name").toLowerCase();
+
+		if (os.contains("win")) {
+			// Windows library name: imgal_c.dll
+			libPrefix = "";
+			libExtension = "dll";
+		} else if (os.contains("mac") || os.contains("darwin")) {
+			// macOS library name: libimgal_c.dylib
+			libPrefix = "lib";
+			libExtension = "dylib";
+		} else {
+			// linux / unix library name: libimgal_c.so
+			libPrefix = "lib";
+			libExtension = "so";
+		}
+
+		// contruct the platform specific library path
+		libPath = "/native/" + libPrefix + libName + "." + libExtension;
+		fileName = libPrefix + libName + "." + libExtension;
+	}
+
+	// copy the imgal library from resources and then load it (for SymbolLookup)
 	static {
 		try {
 			URL url = AbstractNativeLibrary.class.getResource(libPath);
-			Path tmpLib = Files.createTempFile("libimgal_c", "so");
+			if (url == null) {
+				throw new RuntimeException("Native library " + fileName + " not found at: " + libPath);
+			}
+			Path tmpLib = Files.createTempFile(libPrefix + libName, libExtension);
 			try (InputStream is = url.openStream()) {
 				Files.copy(is, tmpLib, StandardCopyOption.REPLACE_EXISTING);
 			}
 			tmpLib.toFile().deleteOnExit();
 			libLookup = SymbolLookup.libraryLookup(tmpLib, Arena.global());
 		} catch (Exception e) {
-			throw new RuntimeException("Failed to load library.", e);
+			throw new RuntimeException("Failed to load library: " + libPath, e);
 		}
 	}
 }
